daemon: split unsupported buy implementation to its own api_*.go files
have a TODO about the fact that all of this could be just made
to return errors
