+++ daemon/api_test.go
@@ -3593,139 +3593,6 @@ func snapList(rawSnaps interface{}) []map[string]interface{} {
 	return snaps
 }
 
-const validBuyInput = `{
-		  "snap-id": "the-snap-id-1234abcd",
-		  "snap-name": "the snap name",
-		  "price": 1.23,
-		  "currency": "EUR"
-		}`
-
-var validBuyOptions = &client.BuyOptions{
-	SnapID:   "the-snap-id-1234abcd",
-	Price:    1.23,
-	Currency: "EUR",
-}
-
-var buyTests = []struct {
-	input                string
-	result               *client.BuyResult
-	err                  error
-	expectedStatus       int
-	expectedResult       interface{}
-	expectedResponseType ResponseType
-	expectedBuyOptions   *client.BuyOptions
-}{
-	{
-		// Success
-		input: validBuyInput,
-		result: &client.BuyResult{
-			State: "Complete",
-		},
-		expectedStatus: 200,
-		expectedResult: &client.BuyResult{
-			State: "Complete",
-		},
-		expectedResponseType: ResponseTypeSync,
-		expectedBuyOptions:   validBuyOptions,
-	},
-	{
-		// Fail with internal error
-		input: `{
-		  "snap-id": "the-snap-id-1234abcd",
-		  "price": 1.23,
-		  "currency": "EUR"
-		}`,
-		err:                  fmt.Errorf("internal error banana"),
-		expectedStatus:       500,
-		expectedResponseType: ResponseTypeError,
-		expectedResult: &errorResult{
-			Message: "internal error banana",
-		},
-		expectedBuyOptions: &client.BuyOptions{
-			SnapID:   "the-snap-id-1234abcd",
-			Price:    1.23,
-			Currency: "EUR",
-		},
-	},
-	{
-		// Fail with unauthenticated error
-		input:                validBuyInput,
-		err:                  store.ErrUnauthenticated,
-		expectedStatus:       400,
-		expectedResponseType: ResponseTypeError,
-		expectedResult: &errorResult{
-			Message: "you need to log in first",
-			Kind:    "login-required",
-		},
-		expectedBuyOptions: validBuyOptions,
-	},
-	{
-		// Fail with TOS not accepted
-		input:                validBuyInput,
-		err:                  store.ErrTOSNotAccepted,
-		expectedStatus:       400,
-		expectedResponseType: ResponseTypeError,
-		expectedResult: &errorResult{
-			Message: "terms of service not accepted",
-			Kind:    "terms-not-accepted",
-		},
-		expectedBuyOptions: validBuyOptions,
-	},
-	{
-		// Fail with no payment methods
-		input:                validBuyInput,
-		err:                  store.ErrNoPaymentMethods,
-		expectedStatus:       400,
-		expectedResponseType: ResponseTypeError,
-		expectedResult: &errorResult{
-			Message: "no payment methods",
-			Kind:    "no-payment-methods",
-		},
-		expectedBuyOptions: validBuyOptions,
-	},
-	{
-		// Fail with payment declined
-		input:                validBuyInput,
-		err:                  store.ErrPaymentDeclined,
-		expectedStatus:       400,
-		expectedResponseType: ResponseTypeError,
-		expectedResult: &errorResult{
-			Message: "payment declined",
-			Kind:    "payment-declined",
-		},
-		expectedBuyOptions: validBuyOptions,
-	},
-}
-
-func (s *apiSuite) TestBuySnap(c *check.C) {
-	s.daemon(c)
-
-	for _, test := range buyTests {
-		s.buyResult = test.result
-		s.err = test.err
-
-		buf := bytes.NewBufferString(test.input)
-		req, err := http.NewRequest("POST", "/v2/buy", buf)
-		c.Assert(err, check.IsNil)
-
-		state := snapCmd.d.overlord.State()
-		state.Lock()
-		user, err := auth.NewUser(state, "username", "email@test.com", "macaroon", []string{"discharge"})
-		state.Unlock()
-		c.Check(err, check.IsNil)
-
-		rsp := postBuy(buyCmd, req, user).(*resp)
-
-		c.Check(rsp.Status, check.Equals, test.expectedStatus)
-		c.Check(rsp.Type, check.Equals, test.expectedResponseType)
-		c.Assert(rsp.Result, check.FitsTypeOf, test.expectedResult)
-		c.Check(rsp.Result, check.DeepEquals, test.expectedResult)
-
-		c.Check(s.buyOptions, check.DeepEquals, test.expectedBuyOptions)
-		c.Check(s.user, check.Equals, user)
-	}
-}
-
 func (s *apiSuite) TestIsTrue(c *check.C) {
 	form := &multipart.Form{}
 	c.Check(isTrue(form, "foo"), check.Equals, false)
@@ -3736,64 +3603,6 @@ func (s *apiSuite) TestIsTrue(c *check.C) {
 	for _, t := range []string{"true", "1", "True", "t"} {
 		form.Value = map[string][]string{"foo": {t}}
 		c.Check(isTrue(form, "foo"), check.Equals, true, check.Commentf("expected %q to be true", t))
-	}
-}
-
-var readyToBuyTests = []struct {
-	input    error
-	status   int
-	respType interface{}
-	response interface{}
-}{
-	{
-		// Success
-		input:    nil,
-		status:   200,
-		respType: ResponseTypeSync,
-		response: true,
-	},
-	{
-		// Not accepted TOS
-		input:    store.ErrTOSNotAccepted,
-		status:   400,
-		respType: ResponseTypeError,
-		response: &errorResult{
-			Message: "terms of service not accepted",
-			Kind:    client.ErrorKindTermsNotAccepted,
-		},
-	},
-	{
-		// No payment methods
-		input:    store.ErrNoPaymentMethods,
-		status:   400,
-		respType: ResponseTypeError,
-		response: &errorResult{
-			Message: "no payment methods",
-			Kind:    client.ErrorKindNoPaymentMethods,
-		},
-	},
-}
-
-func (s *apiSuite) TestReadyToBuy(c *check.C) {
-	s.daemon(c)
-
-	for _, test := range readyToBuyTests {
-		s.err = test.input
-
-		req, err := http.NewRequest("GET", "/v2/buy/ready", nil)
-		c.Assert(err, check.IsNil)
-
-		state := snapCmd.d.overlord.State()
-		state.Lock()
-		user, err := auth.NewUser(state, "username", "email@test.com", "macaroon", []string{"discharge"})
-		state.Unlock()
-		c.Check(err, check.IsNil)
-
-		rsp := readyToBuy(readyToBuyCmd, req, user).(*resp)
-		c.Check(rsp.Status, check.Equals, test.status)
-		c.Check(rsp.Type, check.Equals, test.respType)
-		c.Assert(rsp.Result, check.FitsTypeOf, test.response)
-		c.Check(rsp.Result, check.DeepEquals, test.response)
 	}
 }
